거리 데이터 타입, 거리 계산식 수정

+++ hackathon/src/main/java/org/spring/hackathon/member/constructor/MemberConstructor.java
@@ -60,7 +60,7 @@ public class MemberConstructor {
   }
 
   //플로깅 종료 후 플로깅 한 거리와 얻은 포인트 기록
-  public static MemberEntity memberPloggingEndTransfer(MemberEntity updateMember, float totalDistance, int totalPoint) {
+  public static MemberEntity memberPloggingEndTransfer(MemberEntity updateMember, double totalDistance, int totalPoint) {
 
     updateMember.setPloggingDistanceTotal(totalDistance);
     updateMember.setPloggingPoint(totalPoint);
